+++ test/words.spec.ts
@@ -143,4 +143,18 @@ describe('Normalization of stuffs', () => {
       ['s', ''],
     ]),
   );
+
+  runner.runBasicTests(
+    StringUtilsWord.normalizeSentence,
+    new Map<string, string>([
+      ['        this is a     sentence', 'This is a sentence.'],
+      ['           ', '           '],
+      ['         !!!!!! hello you', 'Hello you.'],
+      ['      a !!!! story', 'A !!!! story.'],
+      ['        This is a story.', 'This is a story.'],
+      ['This is a story !', 'This is a story !'],
+      ['Is this a story ?', 'Is this a story ?'],
+      ['Is this a story ? No', 'Is this a story ? No.'],
+    ]),
+  );
 });